v6/crewai-orchestration/requirements.txt
# LLM Integration
google-generativeai>=0.3.0
langchain-google-genai>=0.0.1
crewai>=0.0.20
crewai-tools>=0.1.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
